Wrap enum AST nodes in structs
Previously, several parts of the AST were represented only by an enum,
e.g. Expr and Stmt. While this is fine for now, it is often useful to
have some metadata that is uniform across all variants of a particular
node, e.g. the type of an expression. To support this, new wrapper
structs have been added for these types of nodes.