Add symbol and type table references to AST
In preparation for semantic analysis, blocks in the AST are now
annotated with references to symbol tables. Additionally, a new Program
struct has been added which will also have a table defining the various
types found in the program.

Currently, both symbol and type tables are stub implementations, but
these will be fleshed out in a future commit.