Add a helper class for chaining RefCell refs
In order to perform proper symbol table lookups, it is necessary to be
able to safely work with references which are dereferenced through an
arbitrary chain of RefCells. There is no way to do this in safe Rust,
since the chain can be arbitrarily long and we must prove to the
compiler that all Refs will live as long as the returned reference
lives.

To allow this to be done, a new helper class called ChainRef has been
added. This class represents a reference obtained through an aribtrarily
long chain of RefCells. Note that this class uses unsafe Rust code,
since it is not possible to prove to borrowck that this is safe.
However, the APIs provided by ChainRef are all verified to be safe.