Add constructor names to the symbol table
In some cases, it's not possible to immediately know what type a given
expression should return. If a constructor call were to appear in such a
context, the current symbol table layout would require looking through
all types to find a matching constructor.

To speed this up, constructor names and their corresponding ids are now
being stored in the symbol table. Note that multiple constructors can
have the same name for different types, so each constructor name must be
mapped to a list of possible constructor ids.