Add Type::Unknown
Previously, expressions were initialized with the type Type::Error.
However, it is necessary to be able to differentiate between an
expression which has already been examined, but was invalid in terms of
types, and an expression which has not yet been examined. This is
important since an expression may be analyzed multiple times due to type
inference and we don't want to emit the same error message multiple
times.

To allow this, Type::Unknown has been added, which represents an
expression that hasn't yet been analyzed and thus has an unknown type.