Improve printing for types
Printing of types has been improved. Firstly, the case of
Type::Unresolved with only one type has been made more clear. This can
arise if the expression is ambiguous such that there are multiple ways
to create a value of the same type. Such an expression is inherently
ambiguous.

Additionally, a pretty printer has been added for types which prints the
name of any data or function types. This should be used when printing
error messages.