Fix inability to define mutually recursive types
The M++ language is supposed to have simultaneous declaration syntax.
This is done to allow mutual recursion, both for types and for
functions. However, this poses a problem for type definitions, since we
must examine all type definitions to determine whether one is valid
because a constructor for one could be written to use any other.

Previously, the simple case of self-recursion was handled by adding a
dummy definition for a data type right before examining that type's
constructors. However, this did not properly handle mutual recursion.
For that, it is necessary to make two passes through the declaration
list: one to add dummy type definitions and one to replace them with the
real type definitions.