Add assignability analysis
Previously, the semantic analyzer was accepting assignment to symbols
even when doing so made no sense. For instance, assigning to a symbol
that represented a function or array.

To correct this, the semantic analyzer will now keep track of whether a
given expression is assignable or not and check this where necessary.