Print the AST for analyze regardless of semantic errors
Previously, the analyze subcommand would only print the annotated AST if
there were no parse/semantic errors detected. This was a problem for
some test cases where parts of them were expected to typecheck properly
while other parts were not.

To rectify this, the analyze subcommand now ignores semantic errors when
deciding whether or not to print the AST.