Allow return statements anywhere in a function
Previously, it was required to place a return statement at the end of a
function and to not have any other return statements in the function.
This was needlessly annoying, especially for recursive functions. To
alleviate this, placing `@feature(return_anywhere)' at the start of the
program will enable the ability to put return statements anywhere in a
function and not put one at the end of a function if it is already
guaranteed to return.