Recognize explicit infinite loops when looking for returns
Previously, a function with an infinite loop that cannot fall through to
the next statement would require an extraneous return statement to be
added. To alleviate this, the return checker will now recognize while
loops with a condition of true and assume they always return. While
these loops could loop indefinitely, this is not an issue since such a
loop will never reach the end of the function.