Implement block expressions
A new type of expression has been added: the block expression. These
expressions act just like a statement block except that they include a
value to be used for the resulting expression.

It is allowable for a block expression to not have a value if it will
always return, in which case it will take on the never type.