Remove the deprecated TypeDefinitionTable::defs field
Previously, the definitions of a TypeDefinitionTable were accessed
directly. This had to be changed when unsafe code was introduced into
TypeDefinitionTable to encapsulate the unsafety.

However, the old defs field of TypeDefinitionTable was temporarily left
while porting code to allow everything to continue compiling, even if it
didn't quite work correctly. However, this hack was supposed to have
been removed before those changes were committed.