Move symbol definitions into a SymbolDefinitionTable
Previously, symbol definitions were being stored inline inside
SymbolTables. However, there was no need for this, and it actually
required a lot of ChainRefs to be used where they were not really
necessary.

A better solution is to use a separate table containing symbol
definitions, similar to how the TypeDefinitionTable has separate
definitions for types.