Implement very simple interprocedural analysis
Previously, the optimizer would fail if any registers were used
nonlocally in another function. This was a combination of a couple of
separate problems: firstly, constant propagation did not properly
understand that the values of these registers can unexpectedly change
during a function call; additionally, dead store elimination did not
understand that a store to a nonlocal variable at the end of a function
was not safe to remove, since the value set in that register can be
observed after the function returned.

All of these problems were resolved by implementing some very simple
IPA, which analyzes which nonlocal registers are read/written by a
function or any functions it can call in turn. Using this information,
the optimizer can correctly deduce what registers may have changed or
whose values may be observed after the function returns.