Fix incorrect constant propagation of arguments/nonlocals
Due to the way global constant propagation was working, argument and
nonlocal registers that were set in a loop body were having their value
propagated to instructions in the loop before the register was set.

This was caused by the global constant propagation algorithm not
properly marking these registers as truly being unknown in the start
node. Instead, the algorithm saw only one path that could initialize the
register and assumed that path must have been taken first.