Fix improper results in copy propagation
Due to an oversight, the code for performing copy propagation was
assuming that the copied register's value was not being changed except
by copies. As a result, it was possible for copy propagation to produce
incorrect IL in certain cases, including the following:

read.i32 $0
copy $1 $0
read.i32 $0
print.i32 $1

The copy propagator would modify this IL to the following, which does
not match the semantics of the original IL:

read.i32 $0
read.i32 $0
print.i32 $0

This code would print the *second* value read instead of the first. This
issue stayed dormant for a long time due to the fact that the IL
generator usually doesn't generate any IL looking like this. That said,
it's possible for such IL to arise from other optimizations and this
issue also affects non-local variable access.

The current fix is not great, but it does produce the correct results.
In the future, this should probably be replaced with a more performant
implementation.